fix score board table timer refactoring
